Fix the naming and extracting function to commonUtils
Renaming some variables and extracting a method to common utils
