+++ lib/common/schemaUtilsCommon.js
@@ -345,5 +345,16 @@ module.exports = {
 
   isKnownType: function(schema) {
     return typeof schemaTypeToJsValidator[schema.type] === 'function';
+  },
+
+  getServersPathVars: function(servers) {
+    return servers.reduce((acc, current) => {
+      let newVarNames = current.hasOwnProperty('variables') ?
+        Object.keys(current.variables).filter((varName) => {
+          return !acc.includes(varName);
+        }) :
+        [];
+      return [...acc, ...newVarNames];
+    }, []);
   }
 };
