+++ lib/schemapack.js
@@ -1,5 +1,4 @@
 'use strict';
-
 // This is the default collection name if one can't be inferred from the OpenAPI spec
 const COLLECTION_NAME = 'Imported from OpenAPI 3.0',
   { getConcreteSchemaUtils } = require('./common/versionUtils.js'),
@@ -24,7 +23,8 @@ const COLLECTION_NAME = 'Imported from OpenAPI 3.0',
   // This provides the base class for
   // errors with the input OpenAPI spec
   OpenApiErr = require('./error.js'),
-  schemaUtils = require('./schemaUtils');
+  schemaUtils = require('./schemaUtils'),
+  { getServersPathVars } = require('./common/schemaUtilsCommon.js');
 
 let path = require('path'),
   concreteUtils,
@@ -476,18 +476,8 @@ class SchemaPack {
               localServers = matchedPath.path.hasOwnProperty('servers') ?
                 matchedPath.path.servers :
                 [],
-              getServersPathVars = (servers) => {
-                return servers.reduce((acc, current) => {
-                  let newVarNames = current.hasOwnProperty('variables') ?
-                    Object.keys(current.variables).filter((varName) => {
-                      return !acc.includes(varName);
-                    }) :
-                    [];
-                  return [...acc, ...newVarNames];
-                }, []);
-              },
               serversPathVars = [...getServersPathVars(localServers), ...getServersPathVars(schema.servers)],
-              isNotALocalServerPathVar = (pathVarName) => {
+              isNotAServerPathVar = (pathVarName) => {
                 return !serversPathVars.includes(pathVarName);
               };
             matchedPath.unmatchedVariablesFromTransaction = [];
@@ -498,7 +488,7 @@ class SchemaPack {
                 let matched = transactionPathVar.key === pathVar.key;
                 if (
                   !matched &&
-                  isNotALocalServerPathVar(transactionPathVar.key) &&
+                  isNotAServerPathVar(transactionPathVar.key) &&
                   !matchedPath.unmatchedVariablesFromTransaction.includes(transactionPathVar.key)
                 ) {
                   matchedPath.unmatchedVariablesFromTransaction.push(transactionPathVar.key);
