+++ test/data/validationData/issues/issue#478/path-variable-does-not-match-collection.json
